requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
brotli==1.2.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
cookiecutter==2.6.0
dj-database-url==3.1.0
django==6.0.1
django-allauth==65.13.1
django-browser-reload==1.21.0
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-htmx==1.27.0
django-mathfilters==1.0.0
django-tailwind==4.4.2
django-widget-tweaks==1.5.0
gunicorn==24.1.1
h11==0.16.0
honcho==2.0.0
idna==3.11
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pyyaml==6.0.3
reportlab==4.4.9
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
whitenoise==6.11.0
